+++ src/comms/data/buff_encoder_data.hpp
@@ -1,0 +1,18 @@
+#pragma once 
+
+#if defined(FIRMWARE)
+#include "comms/data/comms_data.hpp"            // for CommsData
+#elif defined(HIVE)
+#include "modules/comms/data/comms_data.hpp"    // for CommsData
+#endif
+
+#include <stdint.h>                             // uintN_t
+
+/// @brief Structure for the buff encoder sensor.
+struct BuffEncoderData : Comms::CommsData {
+    BuffEncoderData() : CommsData(Comms::TypeLabel::BuffEncoderData, Comms::PhysicalMedium::Ethernet, Comms::Priority::Medium, sizeof(BuffEncoderData)) { }
+    /// Sensor ID.
+    uint8_t id;
+    /// Measured angle.
+    float m_angle;
+};
